Share workflow status visuals
